create TokenPair mixin
