+++ src/views/CreatePair.vue
@@ -77,7 +77,7 @@
           </s-button>
         </div>
       </div>
-      <s-button type="primary" :disabled="!areTokensSelected || isEmptyBalance || isInsufficientBalance || !isAvailable" @click="handleConfirmCreatePair">
+      <s-button type="primary" :disabled="!areTokensSelected || isEmptyBalance || isInsufficientBalance || !isAvailable" @click="openConfirmDialog">
         <template v-if="!areTokensSelected">
           {{ t('exchange.chooseTokens') }}
         </template>
@@ -167,7 +167,7 @@
     <select-token :visible.sync="showSelectFirstTokenDialog" account-assets-only not-null-balance-only :asset="secondToken" @select="setFirstToken" />
     <select-token :visible.sync="showSelectSecondTokenDialog" :asset="firstToken" @select="setSecondToken" />
 
-    <confirm-create-pair :visible.sync="showConfirmCreatePairDialog" @confirm="confirmCreatePair" />
+    <confirm-create-pair :visible.sync="showConfirmDialog" @confirm="confirmCreatePair" />
   </div>
 </template>
 
@@ -176,14 +176,14 @@ import { Component, Mixins, Prop } from 'vue-property-decorator'
 import { Action, Getter } from 'vuex-class'
 import { KnownAssets, KnownSymbols, FPNumber } from '@sora-substrate/util'
 
-import TransactionMixin from '@/components/mixins/TransactionMixin'
 import LoadingMixin from '@/components/mixins/LoadingMixin'
-import InputFormatterMixin from '@/components/mixins/InputFormatterMixin'
+import CreateTokenPairMixin from '@/components/mixins/TokenPairMixin'
 import router, { lazyComponent } from '@/router'
 import { Components, PageNames } from '@/consts'
-import { formatNumber, isNumberValue, isWalletConnected, isMaxButtonAvailable, getMaxValue, isXorAccountAsset, hasInsufficientBalance } from '@/utils'
 
 const namespace = 'createPair'
+
+const TokenPairMixin = CreateTokenPairMixin(namespace)
 
 @Component({
   components: {
@@ -197,90 +197,24 @@ const namespace = 'createPair'
   }
 })
 
-export default class CreatePair extends Mixins(TransactionMixin, LoadingMixin, InputFormatterMixin) {
+export default class CreatePair extends Mixins(TokenPairMixin, LoadingMixin) {
   readonly KnownSymbols = KnownSymbols
 
   @Prop({ type: Boolean, default: false }) readonly parentLoading!: boolean
 
-  @Getter('firstToken', { namespace }) firstToken!: any
-  @Getter('secondToken', { namespace }) secondToken!: any
-  @Getter('firstTokenValue', { namespace }) firstTokenValue!: number
-  @Getter('secondTokenValue', { namespace }) secondTokenValue!: number
-  @Getter('isAvailable', { namespace }) isAvailable!: boolean
   @Getter('minted', { namespace }) minted!: string
-  @Getter('fee', { namespace }) fee!: string
   @Getter('price', { namespace: 'prices' }) price!: string | number
   @Getter('priceReversed', { namespace: 'prices' }) priceReversed!: string | number
 
-  @Action('setFirstToken', { namespace }) setFirstToken
-  @Action('setSecondToken', { namespace }) setSecondToken
-  @Action('getNetworkFee', { namespace }) getNetworkFee
-  @Action('setFirstTokenValue', { namespace }) setFirstTokenValue
-  @Action('setSecondTokenValue', { namespace }) setSecondTokenValue
   @Action('createPair', { namespace }) createPair
   @Action('getPrices', { namespace: 'prices' }) getPrices
   @Action('resetPrices', { namespace: 'prices' }) resetPrices
-
-  showSelectFirstTokenDialog = false
-  showSelectSecondTokenDialog = false
-  inputPlaceholder: string = formatNumber(0, 1)
-  showConfirmCreatePairDialog = false
-  insufficientBalanceTokenSymbol = ''
 
   async mounted () {
     await this.withApi(async () => {
       this.resetPrices()
       await this.setFirstToken(KnownAssets.get(KnownSymbols.XOR))
     })
-  }
-
-  get connected (): boolean {
-    return isWalletConnected()
-  }
-
-  get areTokensSelected (): boolean {
-    return this.firstToken && this.secondToken
-  }
-
-  get isEmptyBalance (): boolean {
-    return +this.firstTokenValue === 0 || +this.secondTokenValue === 0
-  }
-
-  get isFirstMaxButtonAvailable (): boolean {
-    return isMaxButtonAvailable(this.areTokensSelected, this.firstToken, this.firstTokenValue, this.fee)
-  }
-
-  get isSecondMaxButtonAvailable (): boolean {
-    return isMaxButtonAvailable(this.areTokensSelected, this.secondToken, this.secondTokenValue, this.fee)
-  }
-
-  get isInsufficientBalance (): boolean {
-    if (this.connected && this.areTokensSelected) {
-      if (isXorAccountAsset(this.firstToken) || isXorAccountAsset(this.secondToken)) {
-        if (hasInsufficientBalance(this.firstToken, this.firstTokenValue, this.fee)) {
-          this.insufficientBalanceTokenSymbol = this.firstToken.symbol
-          return true
-        }
-        if (hasInsufficientBalance(this.secondToken, this.secondTokenValue, this.fee)) {
-          this.insufficientBalanceTokenSymbol = this.secondToken.symbol
-          return true
-        }
-      }
-      // TODO: Add check for pair without XOR
-    }
-    return false
-  }
-
-  openSelectSecondTokenDialog (): void {
-    this.showSelectSecondTokenDialog = true
-  }
-
-  async handleMaxValue (isFirstToken: boolean): Promise<void> {
-    const token = isFirstToken ? this.firstToken : this.secondToken
-    const action = isFirstToken ? this.setFirstTokenValue : this.setSecondTokenValue
-
-    await this.getNetworkFee()
-    action(getMaxValue(token, this.fee))
   }
 
   updatePrices (): void {
@@ -292,56 +226,8 @@ export default class CreatePair extends Mixins(TransactionMixin, LoadingMixin, I
     })
   }
 
-  async handleTokenChange (isFirstToken: boolean, value: string): Promise<any> {
-    const token = isFirstToken ? this.firstToken : this.secondToken
-    const action = isFirstToken ? this.setFirstTokenValue : this.setSecondTokenValue
-
-    const formattedValue = this.formatNumberField(value, token)
-
-    if (!isNumberValue(formattedValue)) {
-      await this.$nextTick
-      this.resetInputField()
-      return
-    }
-
-    action(formattedValue)
-    this.updatePrices()
-  }
-
-  private resetInputField (isFirstTokenField = true): void {
-    const action = isFirstTokenField ? this.setFirstTokenValue : this.setSecondTokenValue
-    action('')
-  }
-
-  handleInputBlur (isFirstToken: boolean): void {
-    const tokenValue = isFirstToken ? this.firstTokenValue : this.secondTokenValue
-    const action = isFirstToken ? this.setFirstTokenValue : this.setSecondTokenValue
-
-    if (+tokenValue === 0) {
-      this.resetInputField(isFirstToken)
-    } else {
-      const formattedTokenValue = this.trimNeedlesSymbols(String(tokenValue))
-      action(formattedTokenValue)
-    }
-  }
-
-  getTokenBalance (token: any): string {
-    return token?.balance ?? ''
-  }
-
-  handleConfirmCreatePair (): void {
-    this.showConfirmCreatePairDialog = true
-  }
-
-  async confirmCreatePair (): Promise<void> {
-    try {
-      await this.withNotifications(this.createPair)
-      this.showConfirmCreatePairDialog = false
-      router.push({ name: PageNames.Pool })
-    } catch (error) {
-      console.error(error)
-      this.$alert(this.t(error.message), { title: this.t('errorText') })
-    }
+  confirmCreatePair (): Promise<void> {
+    return this.handleConfirm(this.createPair)
   }
 }
 </script>
